+++ src/info.py
@@ -1,4 +1,4 @@
 class Info:
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.11b2" # see https://peps.python.org/pep-0440/
-    AppDateStr = "2023-06-14"
+    AppVersionStr = "0.12b1" # see https://peps.python.org/pep-0440/
+    AppDateStr = "2023-07-12"